Don't let HXCPP_DEBUG_LINK be affected by optimization flags for msvc
Lime and OpenFL typically have a header something like this in files
that should always be compiled for best speed, even in debug mode:

> #if !openfl_debug
> @:fileXml('tags="haxe,release"')
> @:noDebug
> #end

Unfortunately, on Windows, this also prevents debugging information
from going into the pdb file when HXCPP_DEBUG_LINK is passed.

Looking at the other toolchains, it seems like all of the others
simply pass the debugging information flags if HXCPP_DEBUG_LINK is
defined, with no other conditions. This change brings the msvc
toolchain in line with the others.